refactored processes.py into GUI
