+++ main.pyw
@@ -1,19 +1,29 @@
 import sys
 from core.GUI import GUI
-from core.GUI import processes
 
 
 def initialise():
+    """
+    A function that initialises the application
+    :return:
+    """
+    # Creates an application object and set's it's style
     app = GUI.QApplication(sys.argv)
     app.setStyle("Fusion")
 
-    window = processes.ProcessWindow()
+    # creates a MainWindow object and shows it
+    window = GUI.MainWindow()
     window.show()
 
+    # After the window is closed, the program ends
     sys.exit(app.exec())
     
 
 def main():
+    """
+    Starts the program
+    :return:
+    """
     initialise()
     # test
